Rename org.hl7.fhir.convertors package as it's already used
