+++ hapi-fhir-converter/pom.xml
@@ -10,8 +10,7 @@
 	</parent>
 
 	<artifactId>hapi-fhir-converter</artifactId>
-	<!-- TODO: BND changed to jar temporarily -->
-	<packaging>jar</packaging>
+	<packaging>bundle</packaging>
 
 	<dependencies>
 		<dependency>
@@ -152,10 +151,6 @@
 			</resource>
 		</resources>
 		<plugins>
-			<!--
-			TODO: BND
-			With this enabled, JAR file contains a bunch of .class files that
-			seem to get copied in from org.hl7.fhir.convertors
 			<plugin>
 				<groupId>org.apache.felix</groupId>
 				<artifactId>maven-bundle-plugin</artifactId>
@@ -170,7 +165,6 @@
 					</instructions>
 				</configuration>
 			</plugin>
-			-->
 		</plugins>
 	</build>
 </project>
