+++ hapi-fhir-converter/src/test/java/org/hl7/fhir/converter/VersionConvertor_10_30Test.java
@@ -1,7 +1,8 @@
-package org.hl7.fhir.convertors;
+package org.hl7.fhir.converter;
 
 import static org.junit.Assert.assertEquals;
 
+import org.hl7.fhir.convertors.VersionConvertor_10_30;
 import org.hl7.fhir.dstu3.model.*;
 import org.hl7.fhir.exceptions.FHIRException;
 import org.hl7.fhir.dstu2.model.Resource;
